Started migration to Python3
Now that there are python3-qt4 and python3-numpy packages available, and
python3-opengl is in the make, we should start migration of pyFormex to
Python3. As a start we added some tools to help in the migration.
Users should not worry: we will continue support for Python2 for years to
come. But probably will require Python2.6 or Python2.7 as a minimum.