Added a print function check and conversion for scripts
By request of Gianluca: some functionality that you may (not) like.
Scripts are now checked for syntax errors against the print function.
If a print statement is detected, then a conversion is offered, and
if you choose for conversion, the converted script is run. 
This gives an smooth flow in many cases, but is not full-proof: there may
be other errors in the script, or conversion may just fail (e.g. if you
already import print_function in the script). Also, the procedure currently
will read a correct script twice: once to check, once to execute. 
Nevertheless, I though it is useful for running these old scripts that you
forgot to convert, and to help beginning users that are familiar with Python
and are used to the print statement. If it hinders you, switch it off from
the general settings. And offline conversion of ALL your scripts is still 
prefered. E.g., do "2to3 -f print -wn *.py" in all your directories with
pyFormex scripts.