Moved Demo/Horse to HorseStory example; Removed script/app runAll options
The Demos are no longer inccluded in official releases. The Horse demo is 
moved to the normal examples set. The other ones may be removed in future.
The runAll options in the script menus were introduced for testing purposes.
They are not useful for normal users and are moved away from the script menus
to the action menu.