Many small fixes of Examples; deprecated patchextension; added BorderExtrude
This revision contains many small fixes to examples, to get everything running
for the next release. The functionality for creating the Example catalog and
running all examples in sequence has been changed. 
And TriSurface.patchextension is deprecated: most of it can be achieved with
border.extrude (see the new BorderExtrude example) or by a Mesh.connect if
you want a gradual adaption from noncircular to circular section.
Maybe we should allow specification of an alternate end section in the 
extrude?