Fix bug #37833: mesh (deep) copy
Geometry.copy() returns a deep copy, but Mesh.copy() overrided that and
returned a shallow copy. For consistency, we have now renamed Mesh.copy
to Mesh.shallowCopy. I found only one place where Mesh.copy() was used:
in Mesh.levelVolumes(), and replaced it there with shallowCopy.

shallowCopy takes a single argument, prop, to override the
property set in the Mesh. Perhaps the only useful application of creating
a shallow copy is to have the same geometry with another set of properties
on it. Maybe a function 'copy_with_other_props' should become a generic
function of the Geometry class? The current shallowCopy does not allow
you to change the properties to None.