Fixed pyFormex FullVersion after migration to git
utils.FullVersion() has been moved to pyFormex __init__ module (pf.)
FullVersion now shows the stored version and the result of git describe
command in parentheses. Both will be the same if your last commit is the
tag stored as version. On each public release a tag should be created and
stored as the version **BEFORE** creating the tag.