Sanitized tests for integer number
Added arraytools.isInt(obj), to test for a single integer number, whether a
Python int or a numpy integer.
Tests like 'type(obj)==int' or 'isinstance(obj,int)' fail if the object
is not a numpy integer instead of a Python int. They have all been replaced
with isInt(obj).