Add an extra central grid layout in the GUI
The central widget of the pyFormex GUI now has an extra GridLayout
in which the OpenGL viewports are placed. This extra layout level allows
the user to add other widgets to central GUI space.
For example:
  pf.GUI.centralgrid.addWidget(my_widget,0,1)
will add my_widget on row 0, col 1, thus to the right of the viewports.

In future, we may also put the message board inside this grid layout, and
then remove the VBoxLayout currently holding the centralgrid and the message
board.