Change the Table editing
Change the editing of the Table widget (TableModel). This involves a
change of the value setting: the values are no longer eval'ed! Meaning the
table is supposed to store strings, and the user should process these
strings further in needed.

Replace QtGui.QDialog.Accepted with the equivalent ACCEPTED

Also some cleanup of the widgets module.