Fix the app/script configuration/reload from the menu
The app/script paths configuration and the menu reloading were seriously
broken. Both are fixed now. The path configuration allows for direct
editing of the entries. Adding a new path will directly set the text
entry to the (capitalized) basename of the path, so that editing is even
seldomly needed. The save and reload options are gone: on OK, the new
path list is saved and the menu is reloaded. Click Cancel to not save/reload.