Change the use of materials and section databases
The materials and section databases are now global variables of the
properties module.
When creating a PropertyDB, existing MaterialDB and SectionDB or filenames
can be specified. If nothing is given, the defaults from the pyFormex data
directory will be used. The default paths can be changed in the configuration
file.
The global databases can be set by global functions setMaterialDB and
setSectionDB in properties.py, though the corresponding PropertyDB methods
are also retained for compatibility/convenience.