Bump version to 0.9.0a1; change utils.runCommand
The utils.system and utils.runCommand functions have incompatible
changes. The usage of runCommand is now:
runCommand(cmd,timeout=None,RaiseError=True,verbose=True)
and the output is a tuple of (status,stdout,sterr).
Also, the implementation of the timeout has changed, using a timer
(which does not load the processor when running empty).
Please review your code, test and report any problems.

The version change, marking the beginning of the work to a stable 1.0 version,
was also inspired by this incompatible change.