Revert return value of runCommand to sta,out
The return value of the runCommand has been reverted to sta,out only.
Also, the raiseError argument has been removed.
The reason is that runCommand should be as easy as possible for the user.
If an error occurs in the command, runCommand will raise an error, and print
the stderr of the command, so there is no need to return it.
If you really want to get the stderr output you should directly use the
utils.system function, which does return sta,out,err. And if you want to avoid
the error being raised in runCommand, you can always use a try..except (though
in that case you can just directly use utils.system as well).