Fix bug #36805: Display and edit problems for Table widget
Editing of Table widgets can now properly be done inside an InputDialog.
The TableModel and ArrayModel are now distinct classes and provide different
functionality.
ArrayModel is chosen when the input data is a numpy array,
and editing the data will directly change the input data array.
TableModel is chosen when the the input is a list. TableModel can be resized
and get a data type (convertor/validator) per row or column or for the whole
table.

We should probably add an InputTable item to automatically construct the
table widget from the data.