Move conversion of sloppy input items to the InputDialog
Using askItems() allowed for sloppy input of simple items: just a tuple
(key,value) instead of a dict. These items were automatically converted
to a dict. Now this conversion has been moved to the InputDialog itself,
meaning that you can also use the sloppy style when creating a Dialog.
And conversily, it is now also possible to add another widget as an item
in a dialog constructed by askItems(). Formerly this was only possible
when directly using the Dialog class.
So now askItems and Dialog have the same features for the items.