Move CoordinateSystem and BoundVectors
CoordinateSystem and BoundVectors have been removed from the coords.py module
and placed in coordsys.py and plugins/alt.py respectively. There are two
reasons: the main is that they cause the documentation to be cluttered with
a lot of methods inherited from their base class. Most of these methods do
not make any sense on these objects, wihch means that their implementation
probably should be changed (e.g. BoundVectors should be a plex-2 Formex).
And because I want the coords.py to be version 1.0 stable, these changes are
better done in another module.
The coordsys module can get future stuff about coordinate systems.
The plugins/alt is a collector of anything that we do not want to throw
away because somebody needs it and has no place elsewhere.