Add relocation path for moved classes to project.py
When classes are relocated to another module/path, projects saved with
the old class path can not be read back. Adding the path changes to
the project.py module will make the project loader look at the new place
for classes stored with the old path.

Solves support request sr #108239.