Adding filters to expand $Id$ tags
The $Id$ tags are expanded on checkout (by id_smudge) and cleaned
by id_clean on adding to a commit. The filters are installed by running
./install_filters once from the topdir. All users should do this, else
you will get errors on checkout/commit.
Alternatively, you can remove .gitattributes to disable the filtering,
but it will be reinstalled on your next pull.