Add an embedded FileSelection widget.
The FileSelection dialog can now also be embedded in an InputDialog.
The InputFile input item accepts all the arguments of the FileSelection
(except the button, since the dialog buttons are removed anyway, as the
Dialog provides the accept/cancel buttons)
Use:  _I('fn',pf.cfg['workdir'],itemtype='file',text=''),
The text='' removes the label in front of the widget. You may leave it if
you like.