Fix reading Mesh from old project files (bug #38188)
The Mesh __setstate__ has been changed to correctly restore a Mesh that
was saved in older versions: without eltype, with Mesh.eltype (possibly
None), with Mesh.elems.eltype.
In all cases the resulting Mesh has the eltype set in Mesh.elems and
no Mesh.eltype attribute
Remark that the Mesh.eltype is still saved as None. This will be removed
in a future revision, but it does not harm the restoring of the Mesh.

This commit also fixes the setting of the project signature when a read
back project is written out again. It will now correctly store the new
signature instead of the old one.