Add direct read of binary stl files
Finally we have a direct read function for binary STL format:
fileread.read_stl_bin. It is used as a first try by the TriSurface.read
function for all '.stl' files. Even in Python it is quite efficient
(500000 triangles in 3 seconds on my laptop), but we may implement it in
C if someone finds it too slow.

Next, we should also add a binary stl writer.
And we could also think about a direct ascii reader and writer.