Add color to STL export and import
Binary STL files can now include a single color in the header, thus
allowing for storing colored parts. To write a surface with color:
S.write('filename.stl','stlb',color='red')
Any color specification accepted by colors.GLcolor is allowed
When reading a surface with TriSurface.read() from a binary STL with
color, the color will be added to the current viewport's colormap (if it
is not there yet) and the TriSurface gets a matching property number set.

The surface_menu plugin now supports writing colored binary STL, through
the new Export STL option. It provides a dialog to set the file, select
binary or not, and set the color and alpha value. (Beware: the dialog
needs some further work).

The surface_menu now cooperates better with the geometry menu: selecting
(a) surface(s) will set the same selection in the geometry menu.