Add color, alpha and caption attribute detection on Geometry
If a Geometry object is given an attribute alpha, color or caption,
it will be used in the draw() function and in the WebGL export.
Thus you can now write

S.color=red
T.color=blue
draw([S,T])

While this is primarily intended/needed for the WebGL export, it
will also become the natural way of doing things in the new OpenGL
engine of pyFormex (post 1.0 release). But because it was very simple
to implement and we need it for WebGL, I thought it was good to add
the feature already now.