Start conversion to new style signals/slots
In an effort to get up to date with changes in PyQt4 (and to allow
the choice of PySide as an alternative) we started converting the use
of Signals and Slots to the new style. This will go in parallel with
switching everything to the new version 2 API of PyQt4. There is also a
benefit in that the use of signals and slots becomes a lot easier.
Compare the following example. Before:
QtCore.QObject.connect(b,QtCore.SIGNAL("clicked()"),a,QtCore.SLOT("trigger()"))
after:
b.clicked.connect(a.trigger)

The command line version has got an option --pyside, which will force
pyFormex to use PySide (install python-pyside). Beware: there remains a
lot to be converted before everything will work with PySide.