Got pyFormex running on PySide !!
OK, first succes of a full pyFormex GUI running with PySide bindings
instead of PyQt4. Some problems remain to be fixed (many warnings about
cursor shape, MessageBox raises an exception,...), so this is not ready
fo normal use. But users are encouraged to try it out and report bugs
or help fix them.
You need to 'apt-get install python-pyside' and start pyFormex with the
command 'pyformex --pyside'