Make PySide bindings ready for use
Fixed a remaining PyQt4/PySide compatibility problem in imagearray.
Tested PySide on about half of the existing examples.
Added a configuration setting to choose PySide bindings and make it
configurable from Settings Dialog (under the Gui tab).
While we do not intend to make PySide the default in the near future, it
may certainly become so on the longer range.