Make menu.DAction and menu.ActionList work with PySide.
The View and Rendermode buttons, and some parts of the menus, did not work
with PySide, because DACtion and ActionList still used the old signal/slot
mechanism. They have been adapted to new style now, making these items also
work with PySide.

BTW, if you happen to change your default bindings to PySide, you will be
happy to know there is also a --pyqt4 option to force the use of PyQt4
instead of the configured bindings.