Fix saving of images (bug #36995)
Saving the rendering to an image with size different from the current
canvas is now done (again) by resizing the canvas first, grabbing the
image from the current rendering, and then resizing the canvas back to
the original size.
We may come back to offscreen rendering in the future opengl2 engine.