Fix bug #38430; Add configurable InputItem spacing
Two diversions between PySIde and PyQt4 bindings have been resolved.
One about slot usage in addActionButtons (this solves bug 38430).
The other about default spacing of the input items in a dialog. We solved
this by setting the contentsmargin of each inputitem ourself, instead of
relying on the default values. The value is configurable as 'gui/spacing'.