Add PolyLine.movingFrenet
The movingFrenet method computes the Frenet frame in all points of a
PolyLine. This is like the frenet function in nurbs.py, but uses moving
vectors, minimizing the twist. (Thanks, Gianluca).

It is most useful for smooth curves: call it by taking an approx() first.
See Curves example for its use.

Todo:
- for closed curves, the last frame does not coincide with the first:
  the difference in twist should be distributed along the curve.
- the interface should be added as a generic method for all curves (including
  nurbs)