Improved PolyLine.movingFrenet
Some improvements to movingFrenet:
- corrected the docstring
- set the first binormal
- if a binormal fails to be computed, the last one is kept
- add an option to compensate the possible discontinuity at the end for
  closed curves

Note that we do control the *bi*normal, not the normal. The reason is that
the binormal is usually what we want to keep as stable as possible.
Consider a plane curve: the normal to the plane is the binormal, and should
remain constant. By this observation we choose to set the automatic
first binormal as the smallest direction through the first part of the curve.
The user can specify the binormal at the first point. The argument is
currently called upvector, but may be changed in future: there could be
a firstbi and a lastbi, to specify the binormal at bothe ends of an open
curve. That would need some compensation (adding torsion) like we provide
now for closed curves.

If the user wants to specify the normal at the first point, he can compute
the binormal to pass. Or he can pass his normal as upvector and use the
resulting B as normals and -N as binormals.

Testing can be done with the Curves example. Feel free to add interesting
point sets.